+++ src/senaite/core/registry/controlpanel.py
@@ -25,7 +25,9 @@ from plone.registry.interfaces import IRegistry
 from plone.z3cform import layout
 from senaite.core.registry import get_registry_interfaces
 from senaite.core.registry.schema import ISenaiteRegistry
+from senaite.core.interfaces import ISenaiteRegistryFactory
 from zope.component import getUtility
+from zope.component import queryUtility
 from zope.interface import alsoProvides
 
 
@@ -45,7 +47,10 @@ class ContextProxy(object):
 
         registry = getUtility(IRegistry)
         for interface in self.__interfaces:
-            proxy = registry.forInterface(interface)
+            factory = queryUtility(ISenaiteRegistryFactory,
+                                   name=interface.__identifier__,
+                                   default=None)
+            proxy = registry.forInterface(interface, factory=factory)
             try:
                 getattr(proxy, name)
             except AttributeError:
@@ -60,7 +65,10 @@ class ContextProxy(object):
 
         registry = getUtility(IRegistry)
         for interface in self.__interfaces:
-            proxy = registry.forInterface(interface)
+            factory = queryUtility(ISenaiteRegistryFactory,
+                                   name=interface.__identifier__,
+                                   default=None)
+            proxy = registry.forInterface(interface, factory=factory)
             try:
                 return getattr(proxy, name)
             except AttributeError: